requirements.txt
colorama
langchain
langchain_community
pygame
sounddevice
numpy
keyboard
scipy
faster_whisper==1.0.1
kink
gspread
google-auth
google-auth-oauthlib
openai
python-dotenv
inquirer
soundfile
elevenlabs